Fixed the script so the image actualy appeared
Turns out the image wasn't loading so I fixed that problem. It was probably because I declared the "img1" variable twice, instead of just setting the value of the img1 object in the preload function after I declared it at the start of the script (i'm not sure if the vocabulary I'm using is accurate).