Got the second shape to follow the mouse at a delay
I thought I would be able to acess the image's (x,y) directly via img2.x or img2.y but it seems that image-objects have no associated (x,y) positions so I created them myself with the variables img2_x and img2_y.
I then found the difference in the distance between the current position of the image and the position of the mouse, and I then added a fraction of this difference to the image's position so it would slowly "catch up" to the mouse's position.