Added a shape whose position represented a sin function
I created a sin function (sin(x)). I modified it so that its values would be mapped so that they would remain within the dimensions of the canvas while still taking up a sizeable portion of the screen (so you could actually see the representation of the function)
To do this I multiplied sin(x) by the height of the (canvas/4), so the y-values of the function would take up exactly half of the canvas vertically, I then translated the function down by half a screen so the function was centered relative to the canvas. Then I stretched the function horizontally by a factor of 100 so that there would be relatively less waves (periods) within the canvas-space. Finally I iterrated through the xpositions once per frame by 1 and drew a sphere at the x and y point in the function.