Added a incrementing stroke weight to the avatar
Making the avatar alive and subject-like because right now it seems to be too inorganic object-y. This is bad because it intuitively doesn't sense that its moving on its own accord when it looks so lifelss.

Especially because the arrow key movement  does not accelerate or deccelerate (it rather unrealistically sets velocity) the games' entities end up looking a lot colder a life-less and need some more purely aesthetic attention.