Added a fluttering hole in the enemies, using the "noise(){}" function in a better way
Realized that the main reason I wasn't obtaining the desired effect from the noise(){} function is that I was going through its values/perlin-noise-space too quickly.

I then put holes in the spheres which were similarly affected by perlin noise, so there is a nice consistency between the pseudo-random movemnt of the enemies and the fluctuations of their holes.