Gave the Avatar and enemies a motion trail
Drew circles at the positions they would be if the avatar/enemy was moving inversly, this approximates where the avatar/enemy last positions were without having to literally record those positions, therefore giving the appearence of motion trails, emphasising the speed of the avatar/enemies.
Also I ajusted the appearence of the holes inside the enemies, and made them react to the enemy's changing velocity aswell.