Player animation also controlled with perlin noise
I made the player's inside circle pulse in the same way that the enemies' do, but I also gave the noise function two more dimensions so that the velocity of the player (x or y velocity) increases the rapidity of the pulses/animation.

Also I adjusted the values of all functions related to the perlin noise function so it would look better.