ex5 You can now hit a ball with the back of your paddle and have it bounce off realistically.
in exercise 4 collision was largely paddle specefic so that the ball would always end up with a velocity post collision moving it towards the center, this meant that if you hit a ball with the back of your paddle it would teleport to the front of your paddle unrealistically. I finally thought of an easy way to figure out from which side a paddle was hitting a ball based on the paddles' and ball's velocities. Now you can hit a ball with the back of your paddle and have it response realistically.