project 2, created shape and obj classes to organize light-hit-box-collision
I want to create a 2D lighting system, so first to do this I need to define shapes as a series of functions.

So far I'm thinking that lines should be organized into shapes which will act as anchor for their position (possibly later also their rotation) so that lines all representing-- for instance-- a pong-paddle, can be thought of and moved together so each moving obj can have its own associated light-hit-box shape.